requirements.txt
# Core
torch>=2.2.0
lightning>=2.2.0

# Data
pandas>=2.0.0
numpy>=1.26.0
pyarrow>=14.0.0        # parquet support for pandas
scikit-learn>=1.4.0

# ANN index (used in build_index.py and eval)
faiss-cpu>=1.8.0       # swap for faiss-gpu on GPU machines

# Notebook
jupytext>=1.16.0
matplotlib>=3.8.0

# Testing
pytest>=8.0.0